Add SQLite adapter
