+++ tests/Gaufrette/Adapter/MySQLTest.php
@@ -2,147 +2,32 @@
 
 namespace Gaufrette\Adapter;
 
+require_once __DIR__.'/AbstractRDBMSTest.php';
+
 /**
  * @author Markus Bachmann <markus.bachmann@bachi.biz>
  */
-class MySQLTest extends \PHPUnit_Framework_TestCase
+class MySQLTest extends AbstractRDBMSTest
 {
     protected function getAdapter()
     {
-        if (!class_exists('PDO')) {
-            $this->markTestSkipped('PDO class not found');
-        }
+        $pdo = new \PDO('mysql://host=127.0.0.1', 'root', '');
+        $pdo->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
 
-        try {
-            $adapter = new \PDO('mysql://host=localhost;dbname=gaufrette_test', 'root', 'ahb509');
-        } catch (\Exception $e) {
-            echo $e->getMessage();
-            $this->markTestSkipped('Unable to connect to "localhost"');
-        }
+        $pdo->exec('DROP DATABASE `gaufrette_test`');
+        $pdo->exec('CREATE DATABASE `gaufrette_test`');
+        $pdo->exec('USE gaufrette_test');
+        $pdo->exec('CREATE TABLE IF NOT EXISTS files (
+            `id` INTEGER(11) AUTO_INCREMENT PRIMARY KEY,
+            `filename` VARCHAR(255),
+            `binary` BLOB,
+            `metadata` BLOB,
+            `mtime` DATETIME
+        )');
 
-        try {
-            $adapter->query('DELETE FROM files');
-        } catch (\PDOException $e) {
-        }
 
-        return new MySQL($adapter);
-    }
+        $pdo->exec('SET NAMES "UTF8"');
 
-    public function testWriteFile()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertEquals(5, $adapter->write('foo', 'hello'));
-        $this->assertEquals(5, $adapter->write('foo', 'üüüüü'));
-    }
-
-    public function testExistFile()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertFalse($adapter->exists('foo'));
-        $adapter->write('foo', 'bar');
-        $this->assertTrue($adapter->exists('foo'));
-    }
-
-    public function testDelete()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertFalse($adapter->exists('foo'));
-        $adapter->write('foo', 'bar');
-        $this->assertTrue($adapter->exists('foo'));
-        $this->assertTrue($adapter->delete('foo'));
-        $this->assertFalse($adapter->exists('foo'));
-    }
-
-    public function testDeleteNotExistingFile()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertFalse($adapter->exists('foo'));
-        $this->assertFalse($adapter->delete('foo'));
-    }
-
-    public function testMtime()
-    {
-        $adapter = $this->getAdapter();
-        $time = time();
-        $adapter->write('foo', '');
-        $mtime = $adapter->mtime('foo');
-
-        $this->assertInternalType('integer', $mtime);
-        $this->assertGreaterThanOrEqual($time, $mtime);
-    }
-
-    public function testMtimeFromNotExistingFile()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertFalse($adapter->exists('foo'));
-        $this->assertFalse($adapter->mtime('foo'));
-    }
-
-    public function testRename()
-    {
-        $adapter = $this->getAdapter();
-
-        $adapter->write('foo', '');
-        $this->assertTrue($adapter->exists('foo'));
-        $this->assertFalse($adapter->exists('bar'));
-
-        $this->assertTrue($adapter->rename('foo', 'bar'));
-        $this->assertFalse($adapter->exists('foo'));
-        $this->assertTrue($adapter->exists('bar'));
-    }
-
-    public function testRenameNotExistingFile()
-    {
-        $adapter = $this->getAdapter();
-
-        $this->assertFalse($adapter->exists('foo'));
-        $this->assertFalse($adapter->rename('foo', 'bar'));
-    }
-
-    /**
-     * @dataProvider getChecksumData
-     */
-    public function testChecksum($content, $checksum)
-    {
-        $key = 'foo';
-        $adapter = $this->getAdapter();
-
-        $adapter->write($key, $content);
-        $this->assertEquals($checksum, $adapter->checksum($key));
-    }
-
-    public function testKeys()
-    {
-        $adapter = $this->getAdapter();
-        
-        $files = array(
-            'foo' => 'bar',
-            'memo.txt' => 'Lorem',
-            'memo.html' => '<!DOCTYPE><html></html>',
-        );
-
-        foreach ($files as $key => $content) {
-            $adapter->write($key, $content);
-        }
-
-        $keys = $adapter->keys();
-
-        $this->assertEquals(count($files), count($keys));
-        $this->assertEquals(array_keys($files), $keys);
-    }
-
-    public function testSupportMetadata()
-    {
-        $adapter = $this->getAdapter();
-        $this->assertTrue($adapter->supportsMetadata());
-    }
-
-    public function getChecksumData()
-    {
-        return array(
-            array('', 'd41d8cd98f00b204e9800998ecf8427e'),
-            array('foo', 'acbd18db4cc2f85cedef654fccc4a4d8'),
-            array(1, 'c4ca4238a0b923820dcc509a6f75849b')
-        );
+        return new RDBMS($pdo);
     }
 }
